% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adeEffectModels.R
\name{ade_model_full_stop}
\alias{ade_model_full_stop}
\title{Full Stop}
\usage{
ade_model_full_stop(drug_ade_both = FALSE)
}
\arguments{
\item{drug_ade_both}{If true, the prescriptions only stop 
when the ADE takes place when the patient
is prescribed the drug at the same time step}
}
\value{
An ADE model function
}
\description{
Returns a ADE effect model. It stops the prescription of the drug
completely (hence, full stop) when the ADE occurs.
}
