% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adr-effect-models.R
\name{adr_model_full_stop}
\alias{adr_model_full_stop}
\title{The ADR model 'full stop'}
\usage{
adr_model_full_stop(drug_adr_both = FALSE)
}
\arguments{
\item{drug_adr_both}{If \code{TRUE}, the prescriptions only stop
when the ADR takes place when the patient
is prescribed the drug at the same time step
(Default: \code{FALSE})}
}
\value{
An ADR model function
}
\description{
\code{adr_model_full_stop} models the effect of
the appearance of an ADR on the drug prescriptions
of patient. In this case, the appearance of an ADR
stops any future drug prescription. This can, for example,
reflect that the ADR is so severe, that the drug
is never prescribed again.
}
\examples{
drug_history <- c(1, 0, 1, 0, 0)
adr_history <- c(0, 0, 1, 1, 0) 

# returns an ADR model function
adr_model <- adr_model_full_stop(drug_adr_both = FALSE) 
adr_model(adr_history, drug_history)

# -> 0

adr_model <- adr_model_full_stop(drug_adr_both = TRUE) 
adr_model(adr_history, drug_history)
# -> 0 
}
\seealso{
Other ADR effect models: 
\code{\link{adr_model_no_effect}()}
}
\concept{ADR effect models}
