% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adr-effect-models.R
\name{adr_model_no_effect}
\alias{adr_model_no_effect}
\title{The ADR model 'no effect'}
\usage{
adr_model_no_effect()
}
\value{
An ADR model function
}
\description{
\code{adr_model_no_effect} models the effect of
the appearance of an ADR on the drug prescriptions
of patient. In this case, the appearance of an ADR
has no effect on the prescription of a drug.
}
\examples{
drug_history <- c(1, 0, 1, 0, 0)
adr_history <- c(0, 0, 1, 1, 0) 

# returns An ADR model function
adr_model <- adr_model_no_effect() 

adr_model(adr_history, drug_history)
# -> 1
}
\seealso{
Other ADR effect models: 
\code{\link{adr_model_full_stop}()}
}
\concept{ADR effect models}
