% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineFirstPrescription.R
\name{determine_first_prescription}
\alias{determine_first_prescription}
\title{First Prescription}
\usage{
determine_first_prescription(simulation_time, min_chance_drug)
}
\arguments{
\item{simulation_time}{The total number of time points for this patient}

\item{min_chance_drug}{The probability of the drug being prescribed 
when previous prescriptions have no influence 
on the current time point}
}
\value{
A value between 1 and \code{simulation_time}
}
\description{
This function is used for modelling patients 
that are exposed to the drug at least once 
during simulation time.
}
