% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug-history-effect-models.R
\name{drug_model_duration}
\alias{drug_model_duration}
\title{The drug history Model 'duration'}
\usage{
drug_model_duration(duration)
}
\arguments{
\item{duration}{The duration that the drug is prescribed}
}
\value{
A drug history model function
}
\description{
\code{drug_model_duration} returns a
drug history model function. The returned
function models the prescription process such
that the drug is prescribed for a maximum number
of time points, i.e., the \code{duration}
}
\examples{
drug_history <- c(1, 0, 1, 0, 0)

# returns an ADR model function
drug_model <- drug_model_duration(3) 

# apply the model
drug_model(drug_history)
}
\seealso{
Other drug history effect models: 
\code{\link{drug_model_markov_chain}()}
}
\concept{drug history effect models}
