% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug-prescription-effect-models.R
\name{drug_model_markov_chain}
\alias{drug_model_markov_chain}
\title{The Drug History Model 'Markov Chain'}
\usage{
drug_model_markov_chain()
}
\value{
A drug history model function
}
\description{
\code{drug_model_markov_chain} returns a
drug history model function. The returned
function models the drug history time series
as a simple Markov chain model.
}
\examples{
drug_history <- c(1, 0, 1, 0, 0)

# returns an ADR model function
drug_model <- drug_model_markov_chain() 

# apply the model
drug_model(drug_history)
}
\seealso{
Other drug history effect models: 
\code{\link{drug_model_duration}()}
}
\concept{drug history effect models}
