% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePatient.R
\name{generate_patient}
\alias{generate_patient}
\title{Generate a Patient}
\usage{
generate_patient(
  simulation_time = 100,
  risk_function = risk_immediate(),
  prescription_model = prescription_model_markov_chain(),
  ade_model = ade_model_no_effect(),
  min_chance_drug = probability_constant(0.01),
  max_chance_drug = probability_constant(0.5),
  min_chance_ade = probability_constant(0.001),
  max_chance_ade = probability_constant(0.2),
  patient_profile = NULL
)
}
\arguments{
\item{simulation_time}{The total number of time steps}

\item{risk_function}{One of the risk models}

\item{prescription_model}{One of the prescription models}

\item{ade_model}{One of the ADE effect models}

\item{min_chance_drug}{The probability of the drug being prescribed
when the drug history and the ADE history have no effect}

\item{max_chance_drug}{The probability of the ADE when
the drug history and the ADE history have the highest
possible effect}

\item{min_chance}{The probability of the ADE when
the drug history has no effect}

\item{max_chance}{The probability of the ADE when
the drug history has the highest possible effect}

\item{patient}{A list with the covariates of the patient (Default: \code{NULL})}
}
\value{
A \code{patient} object; a list with
\item{\code{drug_prescriptions}}{binary vector of the
drug prescription history}
\item{\code{ade_progression}}{binary vector of the
ade progression}
\item{\code{simulation_time}}{The total number of time steps}
\item{\code{patient_profile}}{A list with patient characteristics}
}
\description{
Generates the drug prescription and
ADE history of an individual patient.
}
\examples{
# create a patient
patient_model <- patient_model_sex(prob_male = 0.5)
(patient_profile <- patient_model())

# choose a risk function
risk_function <- risk_immediate()

# how the drug is prescribed over time
prescription_model <- prescription_model_markov_chain() 

# how the occurence of an ADE affect the prescription of the drug
ade_model = ade_model_no_effect()

# minimum and maximal probabilities for the drug being prescribed
min_chance_drug <- probability_sex(prob_male = .1, prob_female = .2)
max_chance_drug <- probability_sex(prob_male = .5, prob_female = .53)

# minimum and maximal probabilities for the ADE occurring 
min_chance_ade <- probability_sex(prob_male = .01, prob_female = .05)
max_chance_ade <- probability_sex(prob_male = .2, prob_female = .3)

generate_patient(simulation_time = 100, 
                 risk_function, 
                 prescription_model,
                 ade_model, 
                 min_chance_drug,
                 max_chance_drug,
                 min_chance_ade, 
                 max_chance_ade, 
                 patient_profile = patient_profile)                
}
\seealso{
\code{\link{generate_cohort}}
}
