% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientModels.R
\name{patient_model_sex}
\alias{patient_model_sex}
\title{Patient Model 'Sex'}
\usage{
patient_model_sex(prob_male = 0.5)
}
\arguments{
\item{prob_male}{The probability that the
patient is 'male' (Default: \code{0.5})}
}
\value{
A patient model function
}
\description{
A patient model helps to generate the covariates
of the patient, e.g., e.g., sex and age. These
patient descriptions can then be used
in any of the other models: drug prescription
models, ADE effect models etc.
In this case, the patient has only the covariate
`sex'.
}
\examples{
patient_model <- patient_model_sex(prob_male = .4)
patient_model()
# -> list with only item 'sex'
}
\seealso{
Other Patient models: 
\code{\link{patient_model_uninformative}()}
}
\concept{Patient models}
