% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugPrescriptionEffectModels.R
\name{prescription_model_duration}
\alias{prescription_model_duration}
\title{The Drug Prescription Model 'duration'}
\usage{
prescription_model_duration(duration)
}
\arguments{
\item{duration}{The duration that the drug is prescribed}
}
\value{
A drug prescription model function
}
\description{
The drug is prescriped for maximally the 
given duration (number of time steps)
}
\seealso{
Other drug prescription effect models: \code{\link{prescription_model_markov_chain}}
}
\concept{drug prescription effect models}
