% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugPrescriptionEffectModels.R
\name{prescription_model_markov_chain}
\alias{prescription_model_markov_chain}
\title{The Drug Prescription Model 'Markov Chain'}
\usage{
prescription_model_markov_chain()
}
\value{
A drug prescription model function
}
\description{
\code{prescription_model_markov_chain} returns a 
drug prescription model function. The returned
function models the drug prescription time series
as a simple Markov chain model.
}
\examples{
drug_prescriptions <- c(1, 0, 1, 0, 0)

# returns an ADE model function
prescription_model <- prescription_model_markov_chain() 

# apply the model
prescription_model(drug_prescriptions)
}
\seealso{
Other drug prescription effect models: 
\code{\link{prescription_model_duration}()}
}
\concept{drug prescription effect models}
