% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilityModels.R
\name{probability_constant}
\alias{probability_constant}
\title{Probability Model 'Constant'}
\usage{
probability_constant(prob, ...)
}
\arguments{
\item{prob}{The probability}
}
\value{
A probability model function
}
\description{
Probability models define how probabilities 
change with the type of patient. In this case,
the type of patient has no effect on the 
probability.
}
\examples{
probability_model <- probability_constant(prob = 0.2)
probability_model() 
# always returns 0.2 
}
\seealso{
Other Probability models: 
\code{\link{probability_sex}()}
}
\concept{Probability models}
