% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskFunctions.R
\name{risk_immediate}
\alias{risk_immediate}
\title{Risk Model 'Immediate'}
\usage{
risk_immediate()
}
\value{
A risk function
}
\description{
A risk model reflects how the probability of
suffering the ADE changes the history of
drug prescriptions of the patient.
It returns 0 when the drug prescription history
has no effect, and 1 when the patient is at
maximal risk.
In this case, the probability of the ADE
is maximal only when the patient is currently
exposed.
}
\examples{
drug_prescriptions <- c(1, 0, 1, 0, 0)
risk_function <- risk_immediate() 
risk_function(drug_prescriptions) 
}
\seealso{
Other Risk models: 
\code{\link{risk_increase_decrease}()},
\code{\link{risk_long_time_after}()},
\code{\link{risk_no_effect}()},
\code{\link{risk_withdrawal}()}
}
\concept{Risk models}
