% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskFunctions.R
\name{risk_increase_decrease}
\alias{risk_increase_decrease}
\title{Risk Model 'Increase/Decrease'}
\usage{
risk_increase_decrease(peak)
}
\arguments{
\item{peak}{The point at which the risk peaks}
}
\value{
A risk function
}
\description{
A risk model reflects how the probability of 
suffering the ADE changes the history of 
drug prescriptions of the patient. 
It returns 0 when the drug prescription history
has no effect, and 1 when the patient is at 
maximal risk.  
In this case, the risk first increases linearly and 
reaches its peak after \code{peak} time points. Then 
it goes down at the same rate before it is zero again.
}
\seealso{
Other Risk models: \code{\link{risk_immediate}},
  \code{\link{risk_long_time_after}},
  \code{\link{risk_no_effect}},
  \code{\link{risk_withdrawal}}
}
