% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-models.R
\name{risk_model_immediate}
\alias{risk_model_immediate}
\title{Risk Model 'Immediate'}
\usage{
risk_model_immediate()
}
\value{
A risk model
}
\description{
A risk model reflects how the probability of
suffering the ADR changes with the drug exposures
of a patient. It returns \code{0} when the drug
prescription history has no effect, and \code{1}
when the patient is at maximal risk.
In this case, the probability of the ADR
is maximal only when the patient is currently
exposed.
}
\examples{
drug_history <- c(1, 0, 1, 0, 0)
risk_model <- risk_model_immediate() 
risk_model(drug_history) 
}
\seealso{
Other Risk models: 
\code{\link{risk_model_increase_decrease}()},
\code{\link{risk_model_long_time_after}()},
\code{\link{risk_model_no_effect}()},
\code{\link{risk_model_withdrawal}()}
}
\concept{Risk models}
