% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-models.R
\name{risk_model_increase_decrease}
\alias{risk_model_increase_decrease}
\title{Risk Model 'Increase/Decrease'}
\usage{
risk_model_increase_decrease(peak)
}
\arguments{
\item{peak}{The point at which the risk peaks}
}
\value{
A risk model
}
\description{
A risk model reflects how the probability of
suffering the ADR changes with the drug exposures
of a patient. It returns \code{0} when the drug
prescription history has no effect, and \code{1}
when the patient is at maximal risk.
In this case, the risk first increases linearly and
reaches its peak after \code{peak} time points. Then
it goes down at the same rate before it hits zero again.
}
\examples{
drug_history <- c(1, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 1)

risk_model <- risk_model_increase_decrease(peak = 3) 
risk_model(drug_history) 
}
\seealso{
Other Risk models: 
\code{\link{risk_model_immediate}()},
\code{\link{risk_model_long_time_after}()},
\code{\link{risk_model_no_effect}()},
\code{\link{risk_model_withdrawal}()}
}
\concept{Risk models}
