% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-models.R
\name{risk_model_no_effect}
\alias{risk_model_no_effect}
\title{Risk Model 'No Effect'}
\usage{
risk_model_no_effect()
}
\value{
A risk model
}
\description{
A risk model reflects how the probability of
suffering the ADR changes with the drug exposures
of a patient. It returns \code{0} when the drug
prescription history has no effect, and \code{1}
when the patient is at maximal risk.
In this case, the drug has no effect on the ADR
what so ever.
}
\examples{
drug_history <- c(1, 0, 1, 0, 0)
risk_model <- risk_model_no_effect() 
risk_model() 
# -> 0
}
\seealso{
Other Risk models: 
\code{\link{risk_model_immediate}()},
\code{\link{risk_model_increase_decrease}()},
\code{\link{risk_model_long_time_after}()},
\code{\link{risk_model_withdrawal}()}
}
\concept{Risk models}
