% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskFunctions.R
\name{risk_withdrawal}
\alias{risk_withdrawal}
\title{Risk Model 'Withdrawal'}
\usage{
risk_withdrawal(rate)
}
\arguments{
\item{rate}{The rate with which the risk dissipates}
}
\value{
A risk function
}
\description{
A risk model reflects how the probability of 
suffering the ADE changes the history of 
drug prescriptions of the patient. 
It returns 0 when the drug prescription history
has no effect, and 1 when the patient is at 
maximal risk.  
In this case, once the patient is no longer exposed
to the drug, the probability of the ADE peaks 
and then dissipates exponentially.
}
\seealso{
Other Risk models: \code{\link{risk_immediate}},
  \code{\link{risk_increase_decrease}},
  \code{\link{risk_long_time_after}},
  \code{\link{risk_no_effect}}
}
\concept{Risk models}
