% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateADEprogression.R
\name{update_ade_progression}
\alias{update_ade_progression}
\title{Update ADE Progression}
\usage{
update_ade_progression(drug_prescriptions, risk_function, min_chance,
  max_chance, ...)
}
\arguments{
\item{drug_prescriptions}{Binary vector denoting the 
drug prescriptions history}

\item{risk_function}{One of the risk models}

\item{min_chance}{The probability of the ADE when 
the drug history has no effect}

\item{max_chance}{The probability of the ADE when 
the drug history has the highest possible effect}
}
\value{
1 or 0
}
\description{
Returns whether or not the ADE occurs at the next
time point (1) or not (0).
}
\seealso{
Other Update functions: \code{\link{update_drug_prescription}}
}
