% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adeEffectModels.R
\name{ade_model_full_stop}
\alias{ade_model_full_stop}
\title{The ADE model 'full stop'}
\usage{
ade_model_full_stop(drug_ade_both = FALSE)
}
\arguments{
\item{drug_ade_both}{If \code{TRUE}, the prescriptions only stop 
when the ADE takes place when the patient
is prescribed the drug at the same time step (Default: \code{FALSE})}
}
\value{
An ADE model function
}
\description{
\code{ade_model_full_stop} models the effect of 
the appearanceof an ADE on the drug prescriptions 
of patient. In this case, the appearance of an ADE 
stops any future drug prescription. This can, for example, 
reflect that the ADE is so severe, that the drug 
is never prescribed again.
}
\seealso{
Other ADE effect models: 
\code{\link{ade_model_no_effect}()}
}
\concept{ADE effect models}
