% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adeEffectModels.R
\name{ade_model_no_effect}
\alias{ade_model_no_effect}
\title{The ADE model 'no effect'}
\usage{
ade_model_no_effect()
}
\value{
An ADE model function
}
\description{
This function models the effect of the appearance
of an ADE on the drug prescriptions of patient. 
In this case, the appearance of an ADE has no 
effect what so ever on the prescription of a
drug.
}
\seealso{
Other ADE effect models: \code{\link{ade_model_full_stop}}
}
