% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-cohort.R
\name{generate_cohort}
\alias{generate_cohort}
\title{Generate a Cohort}
\usage{
generate_cohort(
  n_patients = 100,
  simulation_time = 30,
  risk_model = risk_model_immediate(),
  drug_model = drug_model_markov_chain(),
  adr_model = adr_model_no_effect(),
  min_chance_drug = probability_constant(0.01),
  max_chance_drug = probability_constant(0.5),
  min_chance_adr = probability_constant(0.001),
  max_chance_adr = probability_constant(0.2),
  patient_model = patient_model_uninformative(),
  verbose = FALSE
)
}
\arguments{
\item{n_patients}{Number of patients}

\item{simulation_time}{The total number of time steps}

\item{risk_model}{One of the risk models}

\item{drug_model}{One of the drug models}

\item{adr_model}{One of the ADR effect models}

\item{min_chance_drug}{The probability of the drug being prescribed
when the drug history and the ADR history have no effect}

\item{max_chance_drug}{The probability of the ADR when
the drug history and the ADR history have the highest
possible effect}
}
\value{
A \code{cohort} object; a list with
\item{\code{n_patients}}{The total number of patients}
\item{\code{simulation_time}}{The total number of time steps}
\item{\code{drug_history}}{A binary matrix with \code{n_patients}
rows and \code{simulation_time} columns}
\item{\code{adr_history}}{A binary matrix with \code{n_patients}
rows and \code{simulation_time} columns}
\item{\code{patient_profile}}{A list with the patient profiles for each patient}
}
\description{
Generates a cohort of patients, each with a
drug prescription and ADR history.
See for more information and details \code{\link{generate_patient}}.
}
\seealso{
\code{\link{generate_patient}}
}
