% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCohort.R
\name{generate_cohort}
\alias{generate_cohort}
\title{Generate a Patient}
\usage{
generate_cohort(n_patients = 1000, simulation_time = 100,
  risk_function = risk_immediate(),
  prescription_model = prescription_model_markov_chain(),
  ade_model = ade_model_no_effect(),
  min_chance_drug = probability_constant(0.01),
  max_chance_drug = probability_constant(0.5),
  min_chance_ade = probability_constant(0.001),
  max_chance_ade = probability_constant(0.2),
  patient_model = patient_model_sex(0.5), verbose = FALSE)
}
\arguments{
\item{n_patients}{Number of patients}

\item{simulation_time}{The total number of time steps}

\item{risk_function}{One of the risk models}

\item{prescription_model}{One of the prescription models}

\item{ade_model}{One of the ADE effect models}

\item{min_chance_drug}{The probability of the drug being prescribed
when the drug history and the ADE history have no effect}

\item{max_chance_drug}{The probability of the ADE when 
the drug history and the ADE history have the highest 
possible effect}
}
\value{
A list with two matrices: the drug prescriptions and the 
        the ADE progression. Each row is a different patient
}
\description{
Generates an individual patient
}
\seealso{
\code{\link{generate_patient}}
}
