% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePatient.R
\name{generate_patient}
\alias{generate_patient}
\title{Generate a Patient}
\usage{
generate_patient(simulation_time = 100,
  risk_function = risk_immediate(),
  prescription_model = prescription_model_markov_chain(),
  ade_model = ade_model_no_effect(),
  min_chance_drug = probability_constant(0.01),
  max_chance_drug = probability_constant(0.5),
  min_chance_ade = probability_constant(0.001),
  max_chance_ade = probability_constant(0.2), patient = NULL)
}
\arguments{
\item{simulation_time}{The total number of time steps}

\item{risk_function}{One of the risk models}

\item{prescription_model}{One of the prescription models}

\item{ade_model}{One of the ADE effect models}

\item{min_chance_drug}{The probability of the drug being prescribed
when the drug history and the ADE history have no effect}

\item{max_chance_drug}{The probability of the ADE when 
the drug history and the ADE history have the highest 
possible effect}

\item{patient}{A list with the covariates of the patient (Default: \code{NULL})}

\item{min_chance}{The probability of the ADE when 
the drug history has no effect}

\item{max_chance}{The probability of the ADE when 
the drug history has the highest possible effect}
}
\value{
A list with a vector of the drug prescriptions and the 
        the ADE progression
}
\description{
Generates an individual patient
}
\seealso{
\code{\link{generate_cohort}}
}
