% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientModels.R
\name{patient_model_uninformative}
\alias{patient_model_uninformative}
\title{Patient Model 'Uniformative'}
\usage{
patient_model_uninformative(...)
}
\value{
A patient model function
}
\description{
A patient model helps to generate the covariates
of the patient, e.g., e.g., sex and age. These
patient descriptions can then be used
in any of the other models: drug prescription
models, ADE effect models etc.
In this case, the patient has no covariates.
All patients are the same.
}
\examples{
patient_model <- patient_model_uninformative() 
patient_model()
# -> empty list. There are no attributes assigned to the patient
}
\seealso{
Other Patient models: 
\code{\link{patient_model_sex}()}
}
\concept{Patient models}
