% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRisk.R
\name{plot_risk}
\alias{plot_risk}
\title{Plot Risk}
\usage{
plot_risk(
  drug_prescriptions,
  risk_function = risk_immediate(),
  title = "",
  ylim = c(0, 1)
)
}
\arguments{
\item{drug_prescriptions}{Binary vector denoting the 
drug prescription history}

\item{risk_function}{One of the risk models}

\item{title}{Title of the plot}

\item{ylim}{The limits of the y-axis: \code{c(ymin, ymax)}}
}
\description{
Plots the risk function's behavior for 
a given drug prescription history
}
