% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugPrescriptionEffectModels.R
\name{prescription_model_duration}
\alias{prescription_model_duration}
\title{The Drug Prescription Model 'duration'}
\usage{
prescription_model_duration(duration)
}
\arguments{
\item{duration}{The duration that the drug is prescribed}
}
\value{
A drug prescription model function
}
\description{
\code{prescription_model_duration} returns a 
drug prescription model function. The returned
function models the prescription process such 
that the drug is prescribed for a maximum number
of time points, i.e., the \code{duration}
}
\examples{
drug_prescriptions <- c(1, 0, 1, 0, 0)

# returns an ADE model function
prescription_model <- prescription_model_duration(3) 

# apply the model
prescription_model(drug_prescriptions)
}
\seealso{
Other drug prescription effect models: 
\code{\link{prescription_model_markov_chain}()}
}
\concept{drug prescription effect models}
