% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugPrescriptionEffectModels.R
\name{prescription_model_markov_chain}
\alias{prescription_model_markov_chain}
\title{The Drug Prescription Model 'Markov Chain'}
\usage{
prescription_model_markov_chain()
}
\value{
A drug prescription model function
}
\description{
The drug prescriptions are modelled as a
Markov Chain.
}
\seealso{
Other drug prescription effect models: \code{\link{prescription_model_duration}}
}
