% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilityModels.R
\name{probability_constant}
\alias{probability_constant}
\title{Probability Model 'Constant'}
\usage{
probability_constant(prob, ...)
}
\arguments{
\item{prob}{The probability}
}
\value{
A probability model function
}
\description{
Probability models define how probabilities 
change with the type of patient. In this case,
the type of patient has no effect on the 
probability.
}
\seealso{
Other Probability models: \code{\link{probability_sex}}
}
