% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskFunctions.R
\name{risk_long_time_after}
\alias{risk_long_time_after}
\title{Risk Model 'Long Time After'}
\usage{
risk_long_time_after(rate, delay)
}
\arguments{
\item{rate}{The rate with which the risk increases}

\item{delay}{How long it takes before the risk is .5}
}
\value{
A risk function
}
\description{
A risk model reflects how the probability of 
suffering the ADE changes the history of 
drug prescriptions of the patient. 
It returns 0 when the drug prescription history
has no effect, and 1 when the patient is at 
maximal risk.  
In this case, the risk of the ADE increases
only after a certian moment in time before it 
reaches the maximal risk.
}
\seealso{
Other Risk models: 
\code{\link{risk_immediate}()},
\code{\link{risk_increase_decrease}()},
\code{\link{risk_no_effect}()},
\code{\link{risk_withdrawal}()}
}
\concept{Risk models}
