% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateADEprogression.R
\name{update_ade_progression}
\alias{update_ade_progression}
\title{Update ADE Progression}
\usage{
update_ade_progression(
  drug_prescriptions,
  risk_function,
  min_chance,
  max_chance,
  patient_profile,
  ...
)
}
\arguments{
\item{drug_prescriptions}{Binary vector denoting the 
drug prescriptions history}

\item{risk_function}{A risk model, e.g., \code{\link{risk_immediate}}}

\item{min_chance}{A probability model function. The probability of the ADE when 
the drug history has no effect}

\item{max_chance}{A probability model function. The probability of the ADE when 
the drug history has the highest possible effect}
}
\value{
\code{1} or \code{0}
}
\description{
Returns whether or not the ADE occurs at the next
time point (\code{1}) or not (\code{0}). The probability of the 
ADE occuring depends on the patient drug prescription
history (\code{drug_prescriptions}) and the risk 
function (\code{risk_function}).
}
\examples{
drug_prescriptions <- c(1, 0, 1, 0, 0)

# choose the risk model
risk_function <- risk_immediate() 

# choose probability models. Note that the probabilities change
# with sex
min_chance <- probability_sex(prob_male = .1, prob_female = .05)
max_chance <- probability_sex(prob_male = .7, prob_female = .8)

# create a patient profile: 
patient_model <- patient_model_sex(prob_male = 0.5) 
(patient_profile <- patient_model()) 

# update the ade_progression
update_ade_progression(drug_prescriptions, 
                       risk_function, 
                       min_chance,
                       max_chance,
                       patient_profile)
}
\seealso{
Other Update functions: 
\code{\link{update_drug_prescription}()}
}
\concept{Update functions}
