% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateDrugPrescription.R
\name{update_drug_prescription}
\alias{update_drug_prescription}
\title{Update Drug Prescription}
\usage{
update_drug_prescription(drug_prescriptions, ade_progression,
  prescription_model, ade_model, min_chance, max_chance, ...)
}
\arguments{
\item{drug_prescriptions}{Binary vector denoting the 
drug prescriptions history}

\item{ade_progression}{Binary vector denoting the history
of the ADE}

\item{prescription_model}{One of the prescription models}

\item{ade_model}{One of the ADE effect models}

\item{min_chance}{The probability of the drug being prescribed
when the drug history and the ADE history have no effect}

\item{max_chance}{The probability of the ADE when 
the drug history and the ADE history have the highest 
possible effect}
}
\value{
1 or 0
}
\description{
Returns whether or not the drug is prescribed 
at the next time point (1) or not (0).
}
\seealso{
Other Update functions: \code{\link{update_ade_progression}}
}
